requirements.txt
flasgger==0.9.7.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
pytest
python-dotenv==1.0.1
marshmallow
Flask-Login
Flask-Security
Flask-JWT-Extended